+++ packages/expect/src/__tests__/matchers.test.js
@@ -455,6 +455,29 @@ describe('.toEqual()', () => {
     expect(actual1).toEqual(expected);
     expect(actual2).not.toEqual(expected);
   });
+
+  test('non-enumerable members should be skipped during equal', () => {
+    const actual = {
+      x: 3,
+    };
+    Object.defineProperty(actual, 'test', {
+      enumerable: false,
+      value: 5,
+    });
+    expect(actual).toEqual({x: 3});
+  });
+
+  test('non-enumerable symbolic members should be skipped during equal', () => {
+    const actual = {
+      x: 3,
+    };
+    const mySymbol = Symbol('test');
+    Object.defineProperty(actual, mySymbol, {
+      enumerable: false,
+      value: 5,
+    });
+    expect(actual).toEqual({x: 3});
+  });
 });
 
 describe('.toBeInstanceOf()', () => {